backend/services/agents/requirements.txt
# FastAPI and ASGI server
fastapi[all]
uvicorn[standard]

# Database
sqlalchemy[asyncio]
asyncpg
alembic

# HTTP client
httpx
aiohttp

# Validation and serialization
pydantic[email]
pydantic-settings

# AI and Agent Frameworks
langchain
langchain-core
langchain-community
crewai
semantic-kernel
autogen-agentchat

# OpenAI and AI providers
openai
anthropic
google-generativeai

# A2A Protocol implementation
websockets
pydantic-extra-types

# Redis for agent state
aioredis

# Observability
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
prometheus-client

# Utilities
python-dotenv
tenacity
structlog

# Development
pytest
pytest-asyncio
black
isort
